\name{flars-package}
\alias{flars-package}
\alias{flars-package}
\docType{package}
\title{
Functional least angle regression for functional linear regression with scalar response and mixed scalar and functional covariates. 
}
\description{
This is a package for the variable selection problem in the functional linear regression model. The model we target on has a scalar response, in other words, a continuous random variable following Normal distribution. The candidate covariates could be either functional or scalar or a mixture of the two. The algorithm is able to do selection when number of candidate variables is larger than the sample size. The efficiency is from the idea of the Least Angle Regression and the stopping rule that we designed for this algorithm. 
}
\details{
\tabular{ll}{
Package: \tab flars\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-05-28\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Yafeng Cheng, Jian Qing Shi

Maintainer: Yafeng Cheng <yafeng.cheng@mrc-bsu.cam.ac.uk>
}
\references{
Cheng, Yafeng, Jian Qing Shi, and Janet Eyre. "Nonlinear Mixed-effects Scalar-on-function Models and Variable Selection for Kinematic Upper Limb Movement Data." arXiv preprint arXiv:1605.06779 (2016).
}
