\name{RealDa}
\alias{RealDa}
\docType{data}
\title{
A subset of real data from Limbs Alive project.
}
\description{
A subset of scalar response variable, functional variables and scalar variables.
}
\usage{RealDa}
\format{
  A list with one response and one list of covariates. The covariates contains 15 functional variables and 11 scalar variables. Each element of the covariates is a matrix.
}

\details{
The data set is from Limbs Alive project. The functional variables are trajectories from patients movements. The scalar variables are summary statistics of some more complex movements and time from stroke to the recording time. Patients' indices are not included in the date set. 
}